LPS-32995 TestUtil class for Layouts
