+++ portal-impl/test/integration/com/liferay/portal/lar/LayoutExportImportTest.java
@@ -28,6 +28,7 @@ import com.liferay.portal.test.EnvironmentExecutionTestListener;
 import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
 import com.liferay.portal.test.TransactionalCallbackAwareExecutionTestListener;
 import com.liferay.portal.util.GroupTestUtil;
+import com.liferay.portal.util.LayoutTestUtil;
 import com.liferay.portlet.sites.util.SitesUtil;
 
 import org.junit.Assert;
@@ -138,7 +139,7 @@ public class LayoutExportImportTest extends BaseExportImportTestCase {
 		throws Exception {
 
 		LayoutSetPrototype layoutSetPrototype =
-			ServiceTestUtil.addLayoutSetPrototype(
+			LayoutTestUtil.addLayoutSetPrototype(
 				ServiceTestUtil.randomString());
 
 		Group layoutSetPrototypeGroup = layoutSetPrototype.getGroup();
@@ -147,10 +148,10 @@ public class LayoutExportImportTest extends BaseExportImportTestCase {
 			LayoutLocalServiceUtil.getLayoutsCount(
 				layoutSetPrototypeGroup, true);
 
-		Layout layoutSetPrototypeLayout1 = ServiceTestUtil.addLayout(
+		Layout layoutSetPrototypeLayout1 = LayoutTestUtil.addLayout(
 			layoutSetPrototypeGroup.getGroupId(),
 			ServiceTestUtil.randomString(), true);
-		Layout layoutSetPrototypeLayout2 = ServiceTestUtil.addLayout(
+		Layout layoutSetPrototypeLayout2 = LayoutTestUtil.addLayout(
 			layoutSetPrototypeGroup.getGroupId(),
 			ServiceTestUtil.randomString(), true);
 
@@ -189,7 +190,7 @@ public class LayoutExportImportTest extends BaseExportImportTestCase {
 
 			if (useLayoutPrototype) {
 				LayoutPrototype layoutPrototype =
-					ServiceTestUtil.addLayoutPrototype(
+					LayoutTestUtil.addLayoutPrototype(
 						ServiceTestUtil.randomString());
 
 				Layout layoutPrototypeLayout = layoutPrototype.getLayout();
@@ -197,7 +198,7 @@ public class LayoutExportImportTest extends BaseExportImportTestCase {
 				updateLayoutTemplateId(layoutPrototypeLayout, "2_2_columns");
 
 				if (layoutPrototypeToLayoutSetPrototype) {
-					Layout layoutSetPrototypeLayout = ServiceTestUtil.addLayout(
+					Layout layoutSetPrototypeLayout = LayoutTestUtil.addLayout(
 						layoutSetPrototypeGroup.getGroupId(),
 						ServiceTestUtil.randomString(), true, layoutPrototype,
 						layoutLinkEnabled);
@@ -209,7 +210,7 @@ public class LayoutExportImportTest extends BaseExportImportTestCase {
 						layoutSetPrototypeLayout.getFriendlyURL());
 				}
 				else {
-					layout = ServiceTestUtil.addLayout(
+					layout = LayoutTestUtil.addLayout(
 						group.getGroupId(), ServiceTestUtil.randomString(),
 						false, layoutPrototype, layoutLinkEnabled);
 				}
@@ -228,7 +229,7 @@ public class LayoutExportImportTest extends BaseExportImportTestCase {
 				}
 			}
 			else {
-				layout = ServiceTestUtil.addLayout(
+				layout = LayoutTestUtil.addLayout(
 					layoutSetPrototypeGroup.getGroupId(),
 					ServiceTestUtil.randomString(), true);
 			}
