+++ portal-impl/test/integration/com/liferay/portal/upgrade/v6_2_0/UpgradeCustomizablePortletsTest.java
@@ -30,6 +30,7 @@ import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
 import com.liferay.portal.test.MainServletExecutionTestListener;
 import com.liferay.portal.test.TransactionalCallbackAwareExecutionTestListener;
 import com.liferay.portal.util.GroupTestUtil;
+import com.liferay.portal.util.LayoutTestUtil;
 import com.liferay.portal.util.PortletKeys;
 import com.liferay.portal.util.TestPropsValues;
 import com.liferay.portlet.PortalPreferencesImpl;
@@ -207,7 +208,7 @@ public class UpgradeCustomizablePortletsTest
 	protected Layout getLayout() throws Exception {
 		Group group = GroupTestUtil.addGroup();
 
-		return ServiceTestUtil.addLayout(
+		return LayoutTestUtil.addLayout(
 			group.getGroupId(), ServiceTestUtil.randomString(), false);
 	}
