+++ portal-impl/test/integration/com/liferay/portal/util/LayoutTestUtil.java
@@ -1,0 +1,110 @@
+/**
+ * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.util;
+
+import com.liferay.portal.NoSuchLayoutException;
+import com.liferay.portal.kernel.util.FriendlyURLNormalizerUtil;
+import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.model.Layout;
+import com.liferay.portal.model.LayoutConstants;
+import com.liferay.portal.model.LayoutPrototype;
+import com.liferay.portal.model.LayoutSetPrototype;
+import com.liferay.portal.service.LayoutLocalServiceUtil;
+import com.liferay.portal.service.LayoutPrototypeLocalServiceUtil;
+import com.liferay.portal.service.LayoutSetPrototypeLocalServiceUtil;
+import com.liferay.portal.service.ServiceContext;
+import com.liferay.portal.service.ServiceTestUtil;
+
+import java.util.HashMap;
+import java.util.Locale;
+
+/**
+ * @author Manuel de la Peña
+ */
+public class LayoutTestUtil {
+
+	public static Layout addLayout(long groupId, String name) throws Exception {
+		return addLayout(groupId, name, false);
+	}
+
+	public static Layout addLayout(
+			long groupId, String name, boolean privateLayout)
+		throws Exception {
+
+		return addLayout(groupId, name, privateLayout, null, false);
+	}
+
+	public static Layout addLayout(
+			long groupId, String name, boolean privateLayout,
+			LayoutPrototype layoutPrototype, boolean linkEnabled)
+		throws Exception {
+
+		String friendlyURL =
+			StringPool.SLASH + FriendlyURLNormalizerUtil.normalize(name);
+
+		Layout layout = null;
+
+		try {
+			layout = LayoutLocalServiceUtil.getFriendlyURLLayout(
+				groupId, false, friendlyURL);
+
+			return layout;
+		}
+		catch (NoSuchLayoutException nsle) {
+		}
+
+		String description = "This is a test page.";
+
+		ServiceContext serviceContext = ServiceTestUtil.getServiceContext();
+
+		if (layoutPrototype != null) {
+			serviceContext.setAttribute(
+				"layoutPrototypeLinkEnabled", linkEnabled);
+			serviceContext.setAttribute(
+				"layoutPrototypeUuid", layoutPrototype.getUuid());
+		}
+
+		return LayoutLocalServiceUtil.addLayout(
+			TestPropsValues.getUserId(), groupId, privateLayout,
+			LayoutConstants.DEFAULT_PARENT_LAYOUT_ID, name, null, description,
+			LayoutConstants.TYPE_PORTLET, false, friendlyURL, serviceContext);
+	}
+
+	public static LayoutPrototype addLayoutPrototype(String name)
+		throws Exception {
+
+		HashMap<Locale, String> nameMap = new HashMap<Locale, String>();
+
+		nameMap.put(LocaleUtil.getDefault(), name);
+
+		return LayoutPrototypeLocalServiceUtil.addLayoutPrototype(
+			TestPropsValues.getUserId(), TestPropsValues.getCompanyId(),
+			nameMap, null, true);
+	}
+
+	public static LayoutSetPrototype addLayoutSetPrototype(String name)
+		throws Exception {
+
+		HashMap<Locale, String> nameMap = new HashMap<Locale, String>();
+
+		nameMap.put(LocaleUtil.getDefault(), name);
+
+		return LayoutSetPrototypeLocalServiceUtil.addLayoutSetPrototype(
+			TestPropsValues.getUserId(), TestPropsValues.getCompanyId(),
+			nameMap, null, true, true, ServiceTestUtil.getServiceContext());
+	}
+
+}
